+++ lib/atcoder/parse_test.go
@@ -1,4 +1,4 @@
-package lib
+package atcoder
 
 import (
 	"github.com/stretchr/testify/assert"
